대소문자/폴더구조/타입분리등 컨벤션수정
