+++ src/components/Modal/Example/TestModal.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Modal from '@components/modal';
+import Modal from '@/components/Modal';
 
 export default function TestModal({
   isOpen,
