+++ src/components/Modal/Header.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import cn from '@lib/cn';
-import { ModalProps } from './types';
+import { ModalProps } from '@/types/modalTypes';
 
 /**
  * @description Modal의 헤더 영역을 정의합니다. `Modal.Title`과 `Modal.Close` 컴포넌트를 포함할 수 있습니다.
